+++ autoload.php
@@ -1,5 +1,7 @@
 <?php
 
+require __DIR__.DIRECTORY_SEPARATOR.'src'.DIRECTORY_SEPARATOR.'functions_include.php';
+
 spl_autoload_register(function ($class) {
 
     $prefix = 'League\Csv\\';