+++ src/functions_include.php
@@ -1,0 +1,6 @@
+<?php
+
+// Don't redefine the functions if included multiple times.
+if (!function_exists('League\Csv\bom_match')) {
+    require __DIR__.'/functions.php';
+}